+++ sdk/dns/Azure.ResourceManager.Dns/src/Customization/Models/DnsSoaRecordData.Serialization.cs
@@ -142,7 +142,6 @@ namespace Azure.ResourceManager.Dns
                 {
                      if (property.Value.ValueKind == JsonValueKind.Null)
                      {
-                         property.ThrowNonNullablePropertyIsNull();
                          continue;
                      }
                      etag = new ETag(property.Value.GetString());
@@ -172,7 +171,6 @@ namespace Azure.ResourceManager.Dns
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
-                        property.ThrowNonNullablePropertyIsNull();
                         continue;
                     }
                     foreach (var property0 in property.Value.EnumerateObject())